refactor[WW-80] : change name package
