+++ src/main/java/com/example/codeE/util/Constant.java
@@ -1,4 +1,4 @@
-package com.example.codeE.ulti;
+package com.example.codeE.util;
 
 //all constant string will be declared here
 public class Constant {
